slack block kit, gitbits, readme-orig
Added a new lab for slack block kit builder. This manifest commands was used in
the slack message on 8 April 2022 to announce the release of ws11 on
nancy's and suresh r's channels.
Added the readme original file
updated fun with gitbits: explained the purpose of git reset for
unstaging all files, and the scenarios of git add all files to staging.